This is the versioin 2.0
The newly added feature of being able to send mass emails with the updated crypto prices. I have also altered it so that the floating point numbers are limited to two decimal places. I personally have this setup on wayscript.com to automatically send me an update of the crypto prices every 4 hours. Not an investor just an enthusiast.